progress on transitioning everything to DFA trait
Tests don't work, but the crate compiles, so we commit for now.

This also adds a new ByteClasses type that encodes some invariants. In
particular, we now put byte classes on the stack to avoid more type
parameters in DenseDFA. In the future, we could make ByteClasses use a
Vec in std environments, but use the stack in no_std environments if
it's an issue.