transition complete!
This is much better and feels much more composable. I'm glad that
DenseDFARef is gone, which I always considered a kludge. One
seeming drawback is that we often need to write monstrosities like
Regex<DenseDFA<Vec<usize>, usize>>>, but I think we can paper over that
with default type parameters and/or aliases.