api: add support for fst::Automata
When the 'fst1' feature is enabled, this will provide impls for the
`fst::Automata` trait for every DFA type (dense and sparse).

Note that this is not intended to be released until the fst 0.4 (or 1.0)
release. Also, the 'fst1' feature is disabled by default since it is
quite niche. (Well, regex-automata is itself niche, but fst support is
probably even more niche!)