nfa: add minimal UTF-8 automata support
This uses Daciuk's algorithm to compile nearly minimal forward UTF-8
automata. For the reverse case, we use a data structure of my own
devising (a "range trie") to organize our Unicode character class such
that it is valid input to Daciuk's algorithm. While this is always
beneficial when constructing a DFA, in the reverse case, it does add
potentially significant overhead to building the NFA. Therefore, we
retain a simpler minimizer based on caching suffixes (like the one found
in the regex crate).

This also re-organizes the `nfa` module into a few different parts, adds
more docs and cleans up some things.