rewrite: the entire crate
There's too much to list in this commit. But the short story is that
I've chosen an intermediate state between where I started and where I
want to be in order to put out a release and unblock some folks. This is
not a release that I'm proud of personally, and plan to keep working on
top of this intermediate state.

In general, unless you have a very specific reason to use this release,
you should wait until 0.3 is out.

Closes #20