beginning the transition in the dfa module
... this is brutal.

On the plus side, going through this lets me see some places where code
might be simplified... but maybe not...

And there are some uses of 'unchecked' that are not obviously correct to
me, although they probably are, and so could use a comment. I left TODO
notes in places.