deps: remove byteorder
This takes #15 and shrinks its diff considerably. Namely, we match the
byteorder API a bit more exactly, remove all changes to
dense/sparse/state_id and keep the minimum things needed in lib.rs and
Cargo.toml.

While improvements are in general welcome, all of this machinery is gone
in the rewrite. The only point of this change is to satisfy a request to
remove a dependency, of which I am sympathetic to. Minimizing the diff
is important for sanely rebasing my rewrite branch on top of this.
Otherwise, removing this dependency probably becomes more annoying than
it's worth.

Closes #15